+++ test/TestPortfolioSwap.t.sol
@@ -17,11 +17,11 @@ contract TestPortfolioSwap is Setup {
         uint128 amtIn = 0.1 ether;
         uint128 amtOut = uint128(subject().getAmountOut(ghost().poolId, sellAsset, amtIn));
 
-        uint prev = ghost().balance(address(this), ghost().quote().to_addr());
+        uint256 prev = ghost().balance(address(this), ghost().quote().to_addr());
         subject().multiprocess(
             EnigmaLib.encodeSwap(uint8(0), ghost().poolId, 0x0, amtIn, 0x0, amtOut, uint8(sellAsset ? 1 : 0))
         );
-        uint post = ghost().balance(address(this), ghost().quote().to_addr());
+        uint256 post = ghost().balance(address(this), ghost().quote().to_addr());
 
         assertTrue(post > prev, "balance-did-not-increase");
     }
